# docs/swagger.yaml
basePath: /api/v1        // Base path for all API routes
definitions:
...
  types.DayPreference:
    enum:
    - any
    - day
    - night
    type: string
    x-enum-comments:
      DayPreferenceAny: No specific preference
      DayPreferenceDay: Primarily daytime activities
      DayPreferenceNight: Primarily evening/night activities
    x-enum-varnames:
    - DayPreferenceAny
    - DayPreferenceDay
    - DayPreferenceNight
...
  types.Response:
    properties:
      error:
        description: Optional error message.
        example: Resource not found
        type: string
      message:
        description: Optional success message.
        example: Operation successful
        type: string
      success:
        description: Indicates if the operation was successful.
        example: true
        type: boolean
    type: object
  types.SearchPace:
    enum:
    - any
    - relaxed
    - moderate
    - fast
    type: string
    x-enum-comments:
      SearchPaceAny: No preference
      SearchPaceFast: Pack in many activities
      SearchPaceModerate: Standard pace
      SearchPaceRelaxed: Fewer, longer activities
    x-enum-varnames:
    - SearchPaceAny
    - SearchPaceRelaxed
    - SearchPaceModerate
    - SearchPaceFast
  types.Tags:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      source:
        type: string
      tag_type:
        description: Consider using a specific enum type
        type: string
      updated_at:
        type: string
    type: object
...
  types.UpdateUserSettingsParams:
    properties:
      default_budget_level:
        type: integer
      default_search_radius_km:
        type: number
      prefer_accessible_pois:
        type: boolean
      prefer_dog_friendly:
        type: boolean
      prefer_outdoor_seating:
        type: boolean
      preferred_pace:
        $ref: '#/definitions/types.SearchPace'
      preferred_time:
        $ref: '#/definitions/types.DayPreference'
    type: object
...
  types.UserSettings:
    properties:
      created_at:
        type: string
      default_budget_level:
        type: integer
      default_search_radius_km:
        description: Use float64 for NUMERIC
        type: number
      prefer_accessible_pois:
        type: boolean
      prefer_dog_friendly:
        type: boolean
      prefer_outdoor_seating:
        type: boolean
      preferred_pace:
        $ref: '#/definitions/types.SearchPace'
      preferred_time:
        $ref: '#/definitions/types.DayPreference'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8000 // Adjust to your actual host/port
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for personalized city discovery and recommendations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WanderWiseAI API
  version: "1.0"
paths:
...
  /user/preferences:
    get:
      consumes:
      - application/json
      description: Retrieves the authenticated user's preferences (profiles).
      produces:
      - application/json
      responses:
        "200":
          description: User Preferences
          schema:
            items:
              $ref: '#/definitions/types.UserSettings'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Get User Preferences
      tags:
      - User
  /user/preferences/{profileID}:
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's preferences (profiles).
      parameters:
      - description: Profile ID
        in: path
        name: profileID
        required: true
        type: string
      - description: Preferences Update Parameters
        in: body
        name: preferences
        required: true
        schema:
          $ref: '#/definitions/types.UpdateUserSettingsParams'
      produces:
      - application/json
      responses:
        "200":
          description: Preferences Updated Successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Update User Preferences
      tags:
      - User
...
  /user/tags:
    get:
      consumes:
      - application/json
      description: Retrieves all tags created by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User Tags
          schema:
            items:
              $ref: '#/definitions/types.Tags'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Get All User Tags
      tags:
      - User Tags
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"